refactor: move download file function to test_util file
