+++ test_util.go
@@ -1,0 +1,37 @@
+package webpbin
+
+import (
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+)
+
+func downloadFile(url, target string) {
+	_, err := os.Stat(target)
+
+	if err != nil {
+		resp, err := http.Get(url)
+
+		if err != nil {
+			fmt.Printf("Error while downloading test image: %v\n", err)
+			panic(err)
+		}
+
+		defer resp.Body.Close()
+
+		f, err := os.Create(target)
+
+		if err != nil {
+			panic(err)
+		}
+
+		defer f.Close()
+
+		_, err = io.Copy(f, resp.Body)
+
+		if err != nil {
+			panic(err)
+		}
+	}
+}
